100% codebase: 0 flake8, 0 mypy (default+strict), 426 tests passing, full docs
- Flake8: 0 errors across 120 source files
- Mypy default: 0 errors
- Mypy strict (--strict): 0 errors (down from 88)
- Pytest: 426 passed

Type fixes:
- vision_tools, pattern_recognition_tools, camera, game_context
- param_panel, boss_venado, scene_registry, drawing_system
- options_scene, stage_scene (12 handler annotations)
- pattern_demo_scene, inventory (TypedDict), world_map_scene
- minimap (Sequence covariance), event_bus (Callable typing)

Documentation:
- CHANGELOG.md, CONTRIBUTING.md
- docs/STAGE_CREATION.md, ENEMY_CREATION.md
- BOSS_CREATION.md, SCENE_CREATION.md

Architecture: EventBus dual pattern already resolved (A1)
Known Gaps: A2-A4 plans documented

+++ src/engine/utils/math_utils.py
@@ -68,7 +68,7 @@ def ease_out_elastic(t: float) -> float:
         return 0.0
     if t > 1.0 - 1e-10:
         return 1.0
-    return 2.0 ** (-10.0 * t) * math.sin((t * 10.0 - 0.75) * 2.094) + 1.0
+    return float(2.0 ** (-10.0 * t) * math.sin((t * 10.0 - 0.75) * 2.094) + 1.0)
 
 
 def ease_in_sine(t: float) -> float:
